worker/requirements.txt
celery==5.3.4
redis==5.0.1
SQLAlchemy==2.0.23
PyMySQL==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.34.0